DesignPatterns/_RealExamDesignPatterns/Exam/SchoolSystem.Framework/Core/Commands/RemoveStudentCommand.cs
using System;
using System.Collections.Generic;
using SchoolSystem.Framework.Core.Commands.Contracts;

namespace SchoolSystem.Framework.Core.Commands
{
    public class RemoveStudentCommand : ICommand
    {
        private ISchoolRegister schoolRegister;

        public RemoveStudentCommand(ISchoolRegister schoolRegister)
        {
            this.schoolRegister = schoolRegister;
        }

        public string Execute(IList<string> parameters)
        {
            var studentId = int.Parse(parameters[0]);
            this.schoolRegister.Students.Remove(studentId);
            return $"Student with ID {studentId} was sucessfully removed.";
        }
    }
}
